// src/shared/enums/garage.enums.ts
export enum TransportTypeEnum {
  car = 'car',
  motorcycle = 'motorcycle',
  custom = 'custom',
}

export enum TransportStyleEnum {
  sportbike = 'sportbike',
  cruisers = 'cruisers',
  standart_naked = 'standart_naked',
  touring_sportTouring = 'touring_sportTouring',
  dualSport = 'dualSport',
  offRoad = 'offRoad',
  electric = 'electric',
}

export enum TransportBodyEnums {
  sedan = 'sedan',
  sportCar = 'sportCar',
  hatchback = 'hatchback',
  muv_suv = 'muv_suv',
  coupe = 'coupe',
  convertible = 'convertible',
  wagon = 'wagon',
  jeep = 'jeep',
  van = 'van',
  minivan = 'minivan',
  crossover = 'crossover',
  landaulet = 'landaulet',
  limousine = 'limousine',
  roadster = 'roadster',
}

export enum TransportTrimEnum {
  CE = 'CE',
  DX = 'DX',
  DL = 'DL',
  EX = 'EX',
  GL = 'GL',
  GLE = 'GLE',
  GT = 'GT',
  LX = 'LX',
  LE = 'LE',
  LS = 'LS',
  LT = 'LT',
  LTD = 'LTD',
  LTZ = 'LTZ',
  SE = 'SE',
  SL = 'SL',
  SLE = 'SLE',
  SLT = 'SLT',
  SV = 'SV',
  XLT = 'XLT',
}

export enum TransportFuelTypeEnum {
  gasoline = 'gasoline',
  electric = 'electric',
  hybrid = 'hybrid',
  diesel = 'diesel',
  ethanol = 'ethanol',
  hydrogen = 'hydrogen',
  biodiesel = 'biodiesel',
  naturalGas = 'naturalGas',
}

export enum TransportHybridEnum {
  plugInHybrid = 'plugInHybrid',
  fullHybrid = 'fullHybrid',
  parallel = 'parallel',
  range = 'range',
}

export enum TransportTransmissionTypeEnum {
  single = 'single',
  automatic = 'automatic',
  automaticManual = 'automaticManual',
  manual = 'manual',
  continuously = 'continuously',
  tiptronic = 'tiptronic',
  semiAutomatic = 'semiAutomatic',
  nonSynchronous = 'nonSynchronous',
  bicycle = 'bicycle',
}

export enum TransportMotorcycleType {
  sportbike = 'sportbike',
  cruisers = 'cruisers',
  standart = 'standart',
  touring = 'touring',
  dualSport = 'dualSport',
  offRoad = 'offRoad',
  electric = 'electric',
}

export enum TransportTypeVehicle {
  motobike = 'Motorbike',
  passenger = 'Passenger',
  largePassager = 'Large Passenger',
  smallTruck = 'Small Track',
  largeTruck = 'Large Truck',
}

export enum TransportColorEnum {
  red = 'red',
  yellow = 'yellow',
  blue = 'blue',
  black = 'black',
  gray = 'gray',
  brown = 'brown',
  white = 'white',
  purple = 'purple',
  silver = 'silver',
  other = 'other',
}

export enum StickerTypeEnum {
  motorbike = 'motorbike',
  passenger = 'passenger',
  largePassenger = 'largePassenger',
  truck = 'truck',
  largeTruck = 'largeTruck',
}

export enum GaragePaymentType {
  sticker = 'sticker',
  license = 'license',
  insurance = 'insurance',
}